GridnetActorHead for microrts; refactor actor
