+++ rl_algo_impls/shared/actor/__init__.py
@@ -1,0 +1,2 @@
+from rl_algo_impls.shared.actor.actor import Actor, PiForward
+from rl_algo_impls.shared.actor.make_actor import actor_head
